+++ cloud/jenkins/psmdb_operator_minikube.groovy
@@ -27,7 +27,21 @@ void runTest(String TEST_NAME) {
             echo Skip $TEST_NAME test
         else
             cd ./source
-            export IMAGE=perconalab/percona-server-mongodb-operator:${env.GIT_BRANCH}
+            if [ -n "${PSMDB_OPERATOR_IMAGE}" ]; then
+                export IMAGE=${PSMDB_OPERATOR_IMAGE}
+            else
+                export IMAGE=perconalab/percona-server-mongodb-operator:${env.GIT_BRANCH}
+            fi
+
+            if [ -n "${IMAGE_MONGOD}" ]; then
+                export IMAGE_MONGOD=${IMAGE_MONGOD}
+            fi
+            if [ -n "${IMAGE_BACKUP}" ]; then
+                export IMAGE_BACKUP=${IMAGE_BACKUP}
+            fi
+            if [ -n "${IMAGE_PMM}" ]; then
+                export IMAGE_PMM=${IMAGE_PMM}
+            fi
 
             ./e2e-tests/$TEST_NAME/run
             touch $FILE_NAME
@@ -58,6 +72,22 @@ pipeline {
             defaultValue: 'https://github.com/percona/percona-server-mongodb-operator',
             description: 'percona-server-mongodb-operator repository',
             name: 'GIT_REPO')
+        string(
+            defaultValue: '',
+            description: 'Operator image: perconalab/percona-server-mongodb-operator:master',
+            name: 'PSMDB_OPERATOR_IMAGE')
+        string(
+            defaultValue: '',
+            description: 'MONGOD image: perconalab/percona-server-mongodb-operator:master-mongod4.0',
+            name: 'IMAGE_MONGOD')
+        string(
+            defaultValue: '',
+            description: 'Backup image: perconalab/percona-server-mongodb-operator:master-backup',
+            name: 'IMAGE_BACKUP')
+        string(
+            defaultValue: '',
+            description: 'PMM image: perconalab/percona-server-mongodb-operator:master-pmm',
+            name: 'IMAGE_PMM')
     }
     agent {
          label 'micro-amazon' 
@@ -93,14 +123,18 @@ pipeline {
                 unstash "sourceFILES"
                 withCredentials([usernamePassword(credentialsId: 'hub.docker.com', passwordVariable: 'PASS', usernameVariable: 'USER')]) {
                     sh '''
-                        cd ./source/
-                        sg docker -c "
-                            docker login -u '${USER}' -p '${PASS}'
-                            export IMAGE=perconalab/percona-server-mongodb-operator:$GIT_BRANCH
-                            ./e2e-tests/build
-                            docker logout
-                        "
-                        sudo rm -rf ./build
+                        if [ -n "${PSMDB_OPERATOR_IMAGE}" ]; then
+                            echo "SKIP: Build is not needed, PSMDB operator image was set!"
+                        else
+                            cd ./source/
+                            sg docker -c "
+                                docker login -u '${USER}' -p '${PASS}'
+                                export IMAGE=perconalab/percona-server-mongodb-operator:$GIT_BRANCH
+                                ./e2e-tests/build
+                                docker logout
+                            "
+                            sudo rm -rf ./build
+                        fi
                     '''
                 }
             }
@@ -122,7 +156,7 @@ pipeline {
 
                         sudo curl -Lo /usr/local/bin/minikube https://storage.googleapis.com/minikube/releases/latest/minikube-linux-amd64
                         sudo chmod +x /usr/local/bin/minikube
-                        minikube start --vm-driver=virtualbox --dns-domain=percona.com --memory=4096 --cpus=3 --kubernetes-version v1.12.8
+                        minikube start --vm-driver=none --dns-domain=percona.com --memory=4096 --cpus=3 --kubernetes-version v1.12.8
                     '''
                     
                     unstash "sourceFILES"
